installer starting from scratch, modern pyproject.toml
