+++ pyproject.toml
@@ -1,0 +1,77 @@
+[build-system]
+requires = ["setuptools >= 63",
+    "setuptools_scm[toml] >= 6.2",
+    "cython >= 0.29",
+    "pkgconfig",
+    ]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "mpf-mc"
+description = "Mission Pinball Framework Media Controller"
+readme = "README.md"
+requires-python = ">=3.7"
+license = {text = "MIT"}
+authors = [{ name = "The Mission Pinball Framework Team", email = "brian@missionpinball.org"}]
+keywords = ['pinball']
+classifiers=['Development Status :: 4 - Beta',
+    'Intended Audience :: Developers',
+    'License :: OSI Approved :: MIT License',
+    'Programming Language :: Python :: 3.7',
+    'Programming Language :: Python :: 3.8',
+    'Programming Language :: Python :: 3.9',
+    'Natural Language :: English',
+    'Operating System :: MacOS :: MacOS X',
+    'Operating System :: Microsoft :: Windows',
+    'Operating System :: POSIX :: Linux',
+    'Topic :: Artistic Software',
+    'Topic :: Games/Entertainment :: Arcade'
+    ]
+dependencies = [
+    "ruamel.yaml == 0.15.100",
+    "mpf >= 0.56.0dev1", # {attr = 'mpfmc._version.__mpf_version_required__'},
+    "kivy ==2.0.0",
+    "psutil ==5.7.3",
+    "Pygments ==2.6.1",
+    "kivy_deps.sdl2 == 0.3.1; platform_system=='Windows'",
+    "kivy_deps.sdl2-dev == 0.3.1; platform_system=='Windows'",
+    "kivy_deps.glew == 0.3.0; platform_system=='Windows'",
+    "kivy_deps.glew-dev == 0.3.0; platform_system=='Windows'",
+    "kivy_deps.gstreamer == 0.3.1; platform_system=='Windows'",
+    "kivy_deps.gstreamer-dev == 0.3.1; platform_system=='Windows'",
+    "ffpyplayer == 4.3.2"
+    ]
+dynamic = ["version"]
+
+[project.urls]
+homepage = "https://missionpinball.org"
+docs = "https://docs.missionpinball.org"
+support-forum = "https://groups.google.com/g/mpf-users"
+
+[project.entry-points."mpf.config_player"]
+sound_player = "mpfmc.config_players.plugins.sound_player:register_with_mpf"
+sound_loop_player = "mpfmc.config_players.plugins.sound_loop_player:register_with_mpf"
+playlist_player = "mpfmc.config_players.plugins.playlist_player:register_with_mpf"
+widget_player = "mpfmc.config_players.plugins.widget_player:register_with_mpf"
+slide_player = "mpfmc.config_players.plugins.slide_player:register_with_mpf"
+track_player = "mpfmc.config_players.plugins.track_player:register_with_mpf"
+display_light_player = "mpfmc.config_players.plugins.display_light_player:register_with_mpf"
+
+[project.entry-points."mpf.command"]
+mc = "mpfmc.commands.mc:get_command"
+imc = "mpfmc.commands.imc:get_command"
+
+[tool.setuptools]
+packages = ["mpfmc"]
+# package-dir = {"" = "mpfmc"}
+include-package-data = true
+
+[tool.setuptools.dynamic]
+version = {attr = "mpfmc._version.__version__"}
+
+[tool.setuptools.package-data]
+mpfmc = [
+    '*.py', '*.yaml', '*.png', '*.md', '*.zip', '*.gif', '*.jpg',
+    '*.mp4', '*.m4v', '*.so', '*.pyd', '*.dylib', '*.wav', '*.ogg',
+    '*.pxd', '*.pyx', '*.c', '*.h', '*.ttf', '*.fnt', '*.txt'
+]
